# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BASELINE Field Operations Intelligence API
  version: 1.0.0
  description: |-
    BASELINE captures operational truth through field visits and derives intelligence from that truth.
    Every endpoint documents its required role, request/response shapes, and applicable audit log action.

    **Role hierarchy**: Admin > Lead > UL (Unit Leader / field worker)
    **Auth**: Bearer token (Azure Entra JWT) on all non-public endpoints.
    **Labor safety**: No endpoint exposes per-worker performance metrics or identifiers
    to Lead/UL roles. `captured_by_oid` is encrypted and Admin-access-only.
  contact:
    name: BASELINE Engineering
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: /api
    description: Production (relative)
...
components:
  securitySchemes:
    AzureAD:
      type: oauth2
      description: 'Azure Active Directory (Entra) OAuth2. Pass the resulting access token as `Authorization: Bearer <token>`.'
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/token
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: Email address
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Resource not found
  responses:
    BadRequest:
      description: Request validation failed — missing or invalid field
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: pool_id is required
    Unauthorized:
      description: Authentication required or Bearer token is invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
    Forbidden:
      description: Caller lacks the required role for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
paths:
...
  /uploads/signed-url:
    post:
      summary: Get a pre-signed S3 upload URL for a stop photo
      description: >
        Returns a time-limited pre-signed S3 URL for direct client-to-S3 upload. The server validates filename and MIME
        type before issuing the URL. The object key is server-generated (UUID-based) — the client filename is never used
        as the storage key. Allowed MIME types: image/jpeg, image/png, image/webp, image/heic.
      tags:
        - Uploads
      security:
        - AzureAD: []
      x-required-roles:
        - UL
        - Lead
        - Admin
      x-audit-action: upload.rejected
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - route_run_stop_id
                - contentType
                - filename
              properties:
                route_run_stop_id:
                  type: integer
                  description: ID of the route run stop this photo belongs to
                  example: 7
                contentType:
                  type: string
                  enum:
                    - image/jpeg
                    - image/png
                    - image/webp
                    - image/heic
                  example: image/jpeg
                filename:
                  type: string
                  description: Original filename (validated, not used as storage key)
                  example: stop_photo.jpg
                kind:
                  type: string
                  enum:
                    - safety
                    - cleaning
                    - infrastructure
                    - completion
                    - skip_after
                  default: completion
                  description: Photo category
            example:
              route_run_stop_id: 7
              contentType: image/jpeg
              filename: stop_photo.jpg
              kind: completion
      responses:
        '200':
          description: Pre-signed URL and server-generated object key
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  uploadUrl:
                    type: string
                    description: Time-limited S3 pre-signed PUT URL
                  objectKey:
                    type: string
                    description: Server-generated S3 object key (use this to reference the photo)
              example:
                ok: true
                uploadUrl: https://s3.amazonaws.com/bucket/runs/7/uuid.jpg?X-Amz-...
                objectKey: runs/7/completion/uuid.jpg
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...